(PC-7565) : move favorite repositories functions to pcapi.core.users
and cleanup unnecessaries indirections
