+++ src/pcapi/infrastructure/container.py
@@ -1,7 +1,6 @@
 from pcapi.infrastructure.repository.beneficiary_bookings.beneficiary_bookings_sql_repository import (
     BeneficiaryBookingsSQLRepository,
 )
-from pcapi.infrastructure.repository.favorite.favorite_sql_repository import FavoriteSQLRepository
 from pcapi.infrastructure.repository.pro_offerers.paginated_offerers_sql_repository import (
     PaginatedOfferersSQLRepository,
 )
@@ -19,12 +18,10 @@ from pcapi.infrastructure.repository.venue.venue_with_offerer_name.venue_with_of
 from pcapi.use_cases.get_bookings_for_beneficiary import GetBookingsForBeneficiary
 from pcapi.use_cases.get_venue_labels import GetVenueLabels
 from pcapi.use_cases.get_venues_by_pro_user import GetVenuesByProUser
-from pcapi.use_cases.list_favorites_of_beneficiary import ListFavoritesOfBeneficiary
 from pcapi.use_cases.list_offerers_for_pro_user import ListOfferersForProUser
 
 
 beneficiary_bookings_repository = BeneficiaryBookingsSQLRepository()
-favorite_repository = FavoriteSQLRepository()
 venue_label_repository = VenueLabelSQLRepository()
 venue_identifier_repository = VenueWithBasicInformationSQLRepository()
 venue_with_offerer_informations_repository = VenueWithOffererNameSQLRepository()
@@ -39,8 +36,6 @@ api_praxiel_stocks = StockProviderPraxielRepository()
 get_venue_labels = GetVenueLabels(venue_label_repository=venue_label_repository)
 
 get_all_venues_by_pro_user = GetVenuesByProUser(venue_repository=venue_with_offerer_informations_repository)
-
-list_favorites_of_beneficiary = ListFavoritesOfBeneficiary(favorite_repository=favorite_repository)
 
 get_bookings_for_beneficiary = GetBookingsForBeneficiary(
     beneficiary_bookings_repository=beneficiary_bookings_repository
