--- src/pcapi/infrastructure/repository/favorite/favorite_sql_repository.py
@@ -1,54 +1,0 @@
-from typing import List
-
-from sqlalchemy.orm import joinedload
-
-from pcapi.domain.favorite.favorite import FavoriteDomain
-from pcapi.domain.favorite.favorite_repository import FavoriteRepository
-from pcapi.infrastructure.repository.favorite import favorite_domain_converter
-from pcapi.models import Booking
-from pcapi.models import Favorite
-from pcapi.models import Offer
-from pcapi.models import Stock
-from pcapi.models import Venue
-
-
-class FavoriteSQLRepository(FavoriteRepository):
-    def find_by_beneficiary(self, beneficiary_identifier: int) -> List[FavoriteDomain]:
-        favorite_sql_entities = (
-            Favorite.query.filter(Favorite.userId == beneficiary_identifier)
-            .options(joinedload(Favorite.offer).joinedload(Offer.venue).joinedload(Venue.managingOfferer))
-            .options(joinedload(Favorite.mediation))
-            .options(joinedload(Favorite.offer).joinedload(Offer.stocks))
-            .options(joinedload(Favorite.offer).joinedload(Offer.product))
-            .options(joinedload(Favorite.offer).joinedload(Offer.mediations))
-            .all()
-        )
-
-        offer_ids = [favorite_sql_entity.offer.id for favorite_sql_entity in favorite_sql_entities]
-
-        bookings = (
-            Offer.query.filter(Offer.id.in_(offer_ids))
-            .join(Stock)
-            .join(Booking)
-            .filter(Booking.userId == beneficiary_identifier)
-            .filter(Booking.isCancelled == False)
-            .with_entities(
-                Booking.id.label("booking_id"),
-                Booking.quantity,
-                Offer.id.label("offer_id"),
-                Stock.id.label("stock_id"),
-            )
-            .all()
-        )
-
-        bookings_by_offer_id = {
-            booking.offer_id: {"id": booking.booking_id, "stock_id": booking.stock_id, "quantity": booking.quantity}
-            for booking in bookings
-        }
-
-        return [
-            favorite_domain_converter.to_domain(
-                favorite_sql_entity, bookings_by_offer_id.get(favorite_sql_entity.offerId, None)
-            )
-            for favorite_sql_entity in favorite_sql_entities
-        ]
