+++ src/pcapi/repository/product_queries.py
@@ -1,6 +1,7 @@
 from typing import Optional
 
 import pcapi.core.offers.repository as offers_repository
+from pcapi.core.users.repository import get_favorites_for_offers
 from pcapi.domain.offers import update_is_active_status
 from pcapi.models import Booking
 from pcapi.models import Offer
@@ -8,7 +9,6 @@ from pcapi.models import Product
 from pcapi.models import Stock
 from pcapi.models import ThingType
 from pcapi.repository import repository
-from pcapi.repository.favorite_queries import get_favorites_for_offers
 from pcapi.repository.mediation_queries import get_mediations_for_offers
 from pcapi.repository.offer_queries import get_offers_by_product_id
